sokoline-backend/requirements.txt
asgiref==3.11.0
boto3==1.42.89
botocore==1.42.89
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dj-database-url==3.1.2
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
jmespath==1.1.0
packaging==26.1
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pycparser==3.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.12.0
